Switch to "Old Brave Dark Theme" name
Brave changed dark themes, so I need to explain which theme is getting
recreated. This is getting bumped to 1.0.0 because the name change could
be considered "breaking" for semantic versioning.

Versioning from now on, should I continue to recreate new theme
revisions, will follow semantic versioning where the major version is
major theme revision, minor version is tweaks to the same revision, and
patches are updating strings in the manifest file.